Update print to logger.info
Also removes logging imports from files that don't do any logging
and removes the example comment. Both of those changes are just minor
non-functional changes to reduce the number of lines of non-functional code.
The logging module is standard to python and there are plenty of examples
in the official Python docs, so I don't think we gain much by including
examples in each file.