filter empty strings from text dataset
When we added `padding='longest'` to the tokenization, it caused empty
strings to be padded to the same length as the longest sample. That
meant that `numel()` was never going to be 0.

This commit moves the empty string check to the part of the code that
loads the dataset into memory.

Because of the way we're sampling indices in `sample_batch`...:

```
        sampled_indices = torch.randperm(len(dataset_split))[:batch_size]
        samples = dataset_split.select(sampled_indices)
        tokenized_outputs = self.text_tokenizer(samples['text'], truncation=True, padding="longest", max_length=self.context_length, return_tensors='pt')
```

... we were ending up with batches of varying sizes, depending on how
many empty text strings we sampled. We always sample 64 indices, for
example, but sometimes we might get 20 of them that are empty strings.
Other times we might get 10 of them that are empty strings. Therefore
our batch size would vary on each sample. By filtering when we load the
dataset into memeory, we guarantee every batch size to be the same.